+++ docs/components/content/examples/vue/tooltip/ExampleVueTooltipRoot.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="grid h-28 place-items-center">
+    <NTooltip
+      content="Hello from tooltip"
+      :_tooltipRoot="{
+        'onUpdate:open': (value: boolean) => {
+          console.log('open', value)
+        },
+      }"
+    >
+      <NButton btn="solid-white">
+        Hover me
+      </NButton>
+    </NTooltip>
+  </div>
+</template>