+++ packages/nuxt/src/runtime/components/elements/tooltip/TooltipTrigger.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+import { TooltipTrigger } from 'radix-vue'
+import type { NTooltipTriggerProps } from '../../../types'
+
+const props = defineProps<NTooltipTriggerProps>()
+</script>
+
+<template>
+  <TooltipTrigger v-bind="props">
+    <slot />
+  </TooltipTrigger>
+</template>