+++ packages/nuxt/src/runtime/types/index.ts
@@ -27,6 +27,7 @@ export * from './table'
 export * from './dropdown-menu'
 export * from './label'
 export * from './popover'
+export * from './tooltip'
 export * from './calendar'
 
 export interface Colors {