+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+# [0.17.0-beta.1](https://github.com/una-ui/una-ui/compare/v0.16.3-beta.1...v0.17.0-beta.1) (2024-09-11)
+
+
+### Features
+
+* **Tooltip:** introduce new component and preset ([#182](https://github.com/una-ui/una-ui/issues/182)) ([fc2a7cc](https://github.com/una-ui/una-ui/commit/fc2a7ccfddda88ee90bca7d0f8e7f3ff9ec77a0d))
+
+
+
+## [0.16.3-beta.1](https://github.com/una-ui/una-ui/compare/v0.16.2-beta.1...v0.16.3-beta.1) (2024-09-11)
+
+
+### Bug Fixes
+
+* **Button:** typo ([df5dc76](https://github.com/una-ui/una-ui/commit/df5dc761943c2fb8c5f670d8ae4d1e424f283172))
+
+
+
+## [0.16.2-beta.1](https://github.com/una-ui/una-ui/compare/v0.16.1-beta.2...v0.16.2-beta.1) (2024-09-11)
+
+
+### Bug Fixes
+
+* **Button:** add bg-transparent to ghost preset ([6f74528](https://github.com/una-ui/una-ui/commit/6f74528e8116ceaec0a7796db1cb48e015888a40))
+
+
+### Features
+
+* **Button:** add transition-colors and remove bg-transparent to base preset ([4648ea6](https://github.com/una-ui/una-ui/commit/4648ea6c63d7d84c47a22500edd96ef929c3cd78))
+
+
+
 ## [0.16.1-beta.2](https://github.com/una-ui/una-ui/compare/v0.16.1-beta.1...v0.16.1-beta.2) (2024-09-10)